Latest WORKING MacPac Script!
This script contains 6 installers it will create for the usb. Edit the code as you please.

To run the script, plug in the usb drive. Only use one in your computer at a time or change the specified disk.
Open the folder in your terminal and run: sudo bash step1.sh
You'll enter your password and it will start to format the drive for 6 mac os installers.
After this script finishes, run: sudo bash step2.sh
Once this is done, check the log in the terminal if all went well. 
Boot your Mac with the flashed USB and hold the option key to enter the partition boot menu.

Enjoy :)